Move e2e helper files to test/e2e/utils dir
