+++ docker/docker-compose.yml
@@ -234,7 +234,7 @@ services:
     extra_hosts:
       - "host.docker.internal:host-gateway"
     volumes:
-      - ../test/e2e/TestControl.php:/var/www/src/Api/Service/TestControl.php
+      - ../test/e2e/utils/TestControl.php:/var/www/src/Api/Service/TestControl.php
 
   test-php:
     build:
