+++ playwright.config.ts
@@ -24,7 +24,7 @@ const config: PlaywrightTestConfig = {
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
   /* Global setup for things like logging in users and saving login cookies */
-  globalSetup: require.resolve('./test/e2e/globalSetup'),
+  globalSetup: require.resolve('./test/e2e/utils/globalSetup'),
   /* Retry on CI only */
   retries: process.env.CI ? 2 : 0,
   /* Opt out of parallel tests on CI. */
